pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mathagent"
version = "0.1.0"
description = "A modular math modeling agent framework scaffold."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

[project.optional-dependencies]
pdf = ["pypdf>=4.0.0"]
ocr = [
  "pymupdf>=1.23.0",
  "pillow>=10.0.0",
  "numpy>=2.0.0",
  "rapidocr-onnxruntime>=1.2.0,<1.3.0",
]
solver = [
  "numpy>=2.0.0",
  "pandas>=2.2.0",
  "scipy>=1.13.0",
  "pulp>=2.8.0",
  "networkx>=3.3",
  "openpyxl>=3.1.5",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
